# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E2BR3 Backend API",
    "description": "OpenAPI documentation for the Axum-based E2BR3 backend. Authenticated API routes use the HTTP-only `auth-token` cookie. Internal machine-to-machine routes use the `x-callback-token` header.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "list_users",
        "parameters": [
          {
            "name": "filters",
            "in": "query",
            "description": "JSON-encoded filters",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "list_options.limit",
            "in": "query",
            "description": "Limit",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "list_options.offset",
            "in": "query",
            "description": "Offset",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "list_options.order_bys",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
    "/api/users/me/password": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "set_my_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMyPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password updated"
          },
          "400": {
            "description": "Invalid password payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "operationId": "update_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserForCreateAdminPayloadDoc"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "SetMyPasswordBodyDoc": {
        "type": "object",
        "required": [
          "new_password"
        ],
        "properties": {
          "new_password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "SetMyPasswordRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SetMyPasswordBodyDoc"
          }
        }
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserForUpdateDoc"
          }
        }
      },
      "UserDoc": {
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "email",
          "username",
          "role",
          "active",
          "must_change_password",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "access_blind_allowed": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "access_end_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_product_ids": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_sender_ids": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_start_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_study_ids": {
            "type": [
              "string",
              "null"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "comments": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "last_login_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "must_change_password": {
            "type": "boolean"
          },
          "organization_id": {
            "type": "string"
          },
          "other_information": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": "string",
            "description": "Canonical role ID. Built-in values include `system_admin`,\n`sponsor_admin_cro`, and `sponsor_admin_company`; other values are\ncustom scoped roles.",
            "example": "sponsor_admin_cro"
          },
          "updated_at": {
            "type": "string"
          },
          "updated_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserForCreateAdminPayloadDoc": {
        "type": "object",
        "required": [
          "organization_id",
          "email"
        ],
        "properties": {
          "access_blind_allowed": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "access_end_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_product_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "access_sender_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "access_start_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_study_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "comments": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "organization_id": {
            "type": "string"
          },
          "other_information": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": [
              "string",
              "null"
            ],
            "description": "Canonical role ID to assign. Use built-in sponsor admin IDs for fixed\nadmin roles, or a custom role name for scoped users.",
            "example": "pvs"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UserForUpdateDoc": {
        "type": "object",
        "properties": {
          "access_blind_allowed": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "access_end_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_product_ids": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_sender_ids": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_start_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "access_study_ids": {
            "type": [
              "string",
              "null"
            ]
          },
          "active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "comments": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_login_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "other_information": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": [
              "string",
              "null"
            ],
            "description": "Canonical role ID. Legacy values are normalized by the backend to the\nnew client-aligned role system.",
            "example": "sponsor_admin_company"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UserListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDoc"
            }
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserDoc"
          }
        }
      }
    },
    "securitySchemes": {
      "auth_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token"
      },
      "callback_token": {
        "type": "apiKey",
        "in": "header",
        "name": "x-callback-token"
      }
    }
  },
...
}